requirements-dev.txt
streamlit>=1.36
pandas>=2.3.2
ruff~=0.14.3
pre-commit~=4.3
pylint>=4,<5
pyright==1.1.407
pytest-cov==7.0.0
coverage>=7.10.6
